refactor: removed external entries creation from outside of the job
