+++ app/Actions/Models/List/ExternalProfile/StoreExternalProfileTokenAction.php
@@ -5,15 +5,12 @@ declare(strict_types=1);
 namespace App\Actions\Models\List\ExternalProfile;
 
 use App\Actions\Http\Api\StoreAction;
-use App\Actions\Models\List\BaseStoreExternalProfileAction;
 use App\Actions\Models\List\ExternalProfile\ExternalEntry\BaseExternalEntryTokenAction;
-use App\Actions\Models\List\ExternalProfile\ExternalEntry\Site\AnilistExternalEntryTokenAction;
+use App\Actions\Models\List\ExternalProfile\ExternalEntry\Token\AnilistExternalEntryTokenAction;
 use App\Enums\Models\List\ExternalProfileSite;
 use App\Enums\Models\List\ExternalProfileVisibility;
-use App\Models\List\External\ExternalEntry;
 use App\Models\List\External\ExternalToken;
 use App\Models\List\ExternalProfile;
-use Error;
 use Exception;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
@@ -22,12 +19,12 @@ use Illuminate\Support\Facades\Log;
 /**
  * Class StoreExternalProfileTokenAction.
  */
-class StoreExternalProfileTokenAction extends BaseStoreExternalProfileAction
+class StoreExternalProfileTokenAction
 {
     protected Collection $resources;
 
     /**
-     * Find or store external profile and its entries given determined external token.
+     * Find or store external profile given determined external token.
      *
      * @param  ExternalToken  $token
      * @param  array  $parameters
@@ -48,31 +45,7 @@ class StoreExternalProfileTokenAction extends BaseStoreExternalProfileAction
 
             $userId = $action->getId();
 
-            // TODO: if the profile already exists, the list should be synced.
-            $profile = $this->searchForUserId($userId, $site, $action, $parameters);
-
-            $entries = $action->getEntries();
-
-            $this->preloadResources($site, $entries);
-
-            $token->externalprofile()->associate($profile);
-
-            $externalEntries = [];
-            foreach ($entries as $entry) {
-                $externalId = Arr::get($entry, 'external_id');
-
-                foreach ($this->getAnimesByExternalId($externalId) as $anime) {
-                    $externalEntries[] = [
-                        ExternalEntry::ATTRIBUTE_SCORE => Arr::get($entry, ExternalEntry::ATTRIBUTE_SCORE),
-                        ExternalEntry::ATTRIBUTE_IS_FAVORITE => Arr::get($entry, ExternalEntry::ATTRIBUTE_IS_FAVORITE),
-                        ExternalEntry::ATTRIBUTE_WATCH_STATUS => Arr::get($entry, ExternalEntry::ATTRIBUTE_WATCH_STATUS),
-                        ExternalEntry::ATTRIBUTE_ANIME => $anime->getKey(),
-                        ExternalEntry::ATTRIBUTE_PROFILE => $profile->getKey(),
-                    ];
-                }
-            }
-
-            ExternalEntry::insert($externalEntries);
+            $profile = $this->findForUserIdOrCreate($userId, $site, $action, $parameters);
 
             return $profile;
 
@@ -90,9 +63,9 @@ class StoreExternalProfileTokenAction extends BaseStoreExternalProfileAction
      * @param  ExternalProfileSite  $site
      * @param  BaseExternalEntryTokenAction  $action
      * @param  array  $parameters
-     * @return ExternalProfile|null
+     * @return ExternalProfile
      */
-    protected function searchForUserId(int $userId, ExternalProfileSite $site, BaseExternalEntryTokenAction $action, array $parameters): ?ExternalProfile
+    protected function findForUserIdOrCreate(int $userId, ExternalProfileSite $site, BaseExternalEntryTokenAction $action, array $parameters): ExternalProfile
     {
         $claimedProfile = ExternalProfile::query()
             ->where(ExternalProfile::ATTRIBUTE_EXTERNAL_USER_ID, $userId)
@@ -122,6 +95,7 @@ class StoreExternalProfileTokenAction extends BaseStoreExternalProfileAction
 
         $storeAction = new StoreAction();
 
+        /** @var ExternalProfile $profile */
         $profile = $storeAction->store(ExternalProfile::query(), [
             ExternalProfile::ATTRIBUTE_EXTERNAL_USER_ID => $userId,
             ExternalProfile::ATTRIBUTE_USER => Arr::get($parameters, ExternalProfile::ATTRIBUTE_USER),
@@ -130,11 +104,7 @@ class StoreExternalProfileTokenAction extends BaseStoreExternalProfileAction
             ExternalProfile::ATTRIBUTE_VISIBILITY => ExternalProfileVisibility::PRIVATE->value,
         ]);
 
-        if ($profile instanceof ExternalProfile) {
-            return $profile;
-        }
-
-        return null;
+        return $profile;
     }
 
     /**
