+++ app/Actions/Models/List/ExternalProfile/StoreExternalProfileUsernameAction.php
@@ -5,12 +5,10 @@ declare(strict_types=1);
 namespace App\Actions\Models\List\ExternalProfile;
 
 use App\Actions\Http\Api\StoreAction;
-use App\Actions\Models\List\BaseStoreExternalProfileAction;
 use App\Actions\Models\List\ExternalProfile\ExternalEntry\BaseExternalEntryAction;
-use App\Actions\Models\List\ExternalProfile\ExternalEntry\Site\AnilistExternalEntryAction;
+use App\Actions\Models\List\ExternalProfile\ExternalEntry\Username\AnilistExternalEntryAction;
 use App\Enums\Models\List\ExternalProfileSite;
 use App\Enums\Models\List\ExternalProfileVisibility;
-use App\Models\List\External\ExternalEntry;
 use App\Models\List\ExternalProfile;
 use Error;
 use Exception;
@@ -22,18 +20,18 @@ use Illuminate\Support\Facades\Log;
 /**
  * Class StoreExternalProfileUsernameAction.
  */
-class StoreExternalProfileUsernameAction extends BaseStoreExternalProfileAction
+class StoreExternalProfileUsernameAction
 {
     /**
      * Find or store an external profile and its entries given determined username.
      *
      * @param  Builder  $builder
      * @param  array  $profileParameters
-     * @return ExternalProfile|null
+     * @return ExternalProfile
      *
      * @throws Exception
      */
-    public function findOrCreate(Builder $builder, array $profileParameters): ?ExternalProfile
+    public function findOrCreate(Builder $builder, array $profileParameters): ExternalProfile
     {
         try {
             $profileSite = ExternalProfileSite::fromLocalizedName(Arr::get($profileParameters, 'site'));
@@ -52,12 +50,8 @@ class StoreExternalProfileUsernameAction extends BaseStoreExternalProfileAction
             $action = $this->getActionClass($profileSite, $profileParameters);
 
             if ($action === null) {
-                return null;
+                throw new Error("Action not found for site {$profileSite->localize()}", 404);
             }
-
-            $entries = $action->getEntries();
-
-            $this->preloadResources($profileSite, $entries);
 
             $storeAction = new StoreAction();
 
@@ -69,23 +63,6 @@ class StoreExternalProfileUsernameAction extends BaseStoreExternalProfileAction
                 ExternalProfile::ATTRIBUTE_VISIBILITY => ExternalProfileVisibility::fromLocalizedName(Arr::get($profileParameters, ExternalProfile::ATTRIBUTE_VISIBILITY))->value,
             ]);
 
-            $externalEntries = [];
-            foreach ($entries as $entry) {
-                $externalId = Arr::get($entry, 'external_id');
-
-                foreach ($this->getAnimesByExternalId($externalId) as $anime) {
-                    $externalEntries[] = [
-                        ExternalEntry::ATTRIBUTE_SCORE => Arr::get($entry, ExternalEntry::ATTRIBUTE_SCORE),
-                        ExternalEntry::ATTRIBUTE_IS_FAVORITE => Arr::get($entry, ExternalEntry::ATTRIBUTE_IS_FAVORITE),
-                        ExternalEntry::ATTRIBUTE_WATCH_STATUS => Arr::get($entry, ExternalEntry::ATTRIBUTE_WATCH_STATUS),
-                        ExternalEntry::ATTRIBUTE_ANIME => $anime->getKey(),
-                        ExternalEntry::ATTRIBUTE_PROFILE => $profile->getKey(),
-                    ];
-                }
-            }
-
-            ExternalEntry::insert($externalEntries);
-
             DB::commit();
 
             return $profile;
@@ -94,7 +71,7 @@ class StoreExternalProfileUsernameAction extends BaseStoreExternalProfileAction
 
             DB::rollBack();
 
-            return null;
+            throw $e;
         }
     }
